pyproject.toml
[project]
name = "deep-researcher"
version = "0.1.0"
description = "Deep Researcher Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.5.0",
    "openai>=1.65.5",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tiktoken>=0.9.0",
]
